refactor(desktop): isolate ACP backend adapter
